역무원 호출 완료 페이지 구현, 호출 상세정보 컴포넌트화
